[FEAT] : moved CreateRoomButton related code to its tsx file from Rooms.tsx
